move header footer to components
